Combo of little features and bug fixes for stability

+++ src/modules/apps.js
@@ -50,15 +50,6 @@ const registryClient = () => new RegistryClient({
   userAgent: userAgent,
 })
 
-const handleError = (err) => {
-  if (err && err.name === 'StatusCodeError' && err.error.exception) {
-    const error = new Error(err.error.exception.message)
-    error.name = err.error.code
-    throw error
-  }
-  throw err
-}
-
 const sendChanges = (() => {
   let queue = []
   const publishPatch = debounce(
@@ -80,7 +71,7 @@ const sendChanges = (() => {
       .catch(err => {
         timeEnd()
         stopSpinner()
-        handleError(err)
+        return Promise.reject(err)
       })
     },
     200
@@ -95,6 +86,7 @@ const sendChanges = (() => {
 })()
 
 const keepAppAlive = () => {
+  let exitPromise
   const rcApp = `${id}+rc`
   return installApp(rcApp)
   .then(() => {
@@ -109,10 +101,13 @@ const keepAppAlive = () => {
       input: process.stdin,
       output: process.stdout,
     }).on('SIGINT', () => {
+      if (exitPromise) {
+        return
+      }
       stopSpinner()
       clearTimeout(keepAliveInterval)
       log.info('Exiting...')
-      appsClient().uninstallApp(getAccount(), getWorkspace(), rcApp)
+      exitPromise = appsClient().uninstallApp(getAccount(), getWorkspace(), rcApp)
       .finally(() => process.exit())
     })
   })
@@ -137,14 +132,24 @@ const publishApp = (file, pre = false) => {
   )
 }
 
+const promptAppUninstall = (app) => {
+  return inquirer.prompt({
+    type: 'confirm',
+    name: 'confirm',
+    message: `Are you sure you want to uninstall the app ${app}?`,
+  })
+  .then(({confirm}) => confirm)
+}
+
+const workspaceMasterMessage = `${chalk.green('master')} is ${chalk.red('read-only')}, please use another workspace`
+
 export default {
   list: {
     alias: 'ls',
-    'optionalArgs': 'query',
+    optionalArgs: 'query',
     description: 'List your installed VTEX apps',
     handler: () => {
       log.debug('Starting to list apps')
-
       return appsClient().listApps(
         getAccount(),
         getWorkspace()
@@ -170,12 +175,19 @@ export default {
   watch: {
     description: 'Send the files to the registry and watch for changes',
     handler: () => {
+      const workspace = getWorkspace()
+      if (workspace === 'master') {
+        log.error(workspaceMasterMessage)
+        return Promise.resolve()
+      }
+
+      const account = getAccount()
       log.info('Watching app', `${id}`)
       console.log(
         chalk.green('Your URL:'),
-        chalk.blue(getWorkspaceURL(getAccount(), getWorkspace()))
+        chalk.blue(getWorkspaceURL(account, workspace))
       )
-      courier.listen(getAccount(), getWorkspace(), getToken())
+      courier.listen(account, workspace, getToken())
       if (log.level === 'info') {
         setSpinnerText('Sending files')
         startSpinner()
@@ -196,19 +208,29 @@ export default {
       .then(() => watch(root, sendChanges))
       .catch(err => {
         stopSpinner()
-        handleError(err)
+        return Promise.reject(err)
       })
     },
   },
   install: {
-    requiredArgs: 'app',
+    optionalArgs: 'app',
     alias: 'i',
-    description: 'Install the specified app',
-    handler: (app) => {
+    description: 'Install an app on the current directory or a specified one',
+    handler: (optionalApp) => {
+      const workspace = getWorkspace()
+      if (workspace === 'master') {
+        log.error(workspaceMasterMessage)
+        return Promise.resolve()
+      }
+
+      const app = typeof optionalApp === 'string'
+        ? optionalApp
+        : `${manifest.vendor}.${manifest.name}@${manifest.version}`
       log.debug('Starting to install app', app)
       const appRegex = new RegExp(`^${vendorPattern}\.${namePattern}@${wildVersionPattern}$`)
       if (!appRegex.test(app)) {
-        return log.error('Invalid app format, please use <vendor>.<name>@<version>')
+        log.error('Invalid app format, please use <vendor>.<name>@<version>')
+        return Promise.resolve()
       }
 
       return installApp(app)
@@ -224,7 +246,21 @@ export default {
   uninstall: {
     requiredArgs: 'app',
     description: 'Uninstall the specified app',
-    handler: (app) => {
+    options: [
+      {
+        short: 'y',
+        long: 'yes',
+        description: 'Auto confirm prompts',
+        type: 'boolean',
+      },
+    ],
+    handler: (app, options) => {
+      const workspace = getWorkspace()
+      if (workspace === 'master') {
+        log.error(workspaceMasterMessage)
+        return Promise.resolve()
+      }
+
       log.debug('Starting to uninstall app', app)
       const appRegex = new RegExp(`^${vendorPattern}\.${namePattern}$`)
       if (!appRegex.test(app)) {
@@ -232,14 +268,9 @@ export default {
         return Promise.resolve()
       }
 
-      return Promise.try(() =>
-        inquirer.prompt({
-          type: 'confirm',
-          name: 'confirm',
-          message: `Are you sure you want to uninstall the app ${app}?`,
-        })
-      )
-      .then(({confirm}) => confirm || Promise.reject('User cancelled'))
+      const preConfirm = options.y || options.yes
+      return Promise.try(() => preConfirm || promptAppUninstall(app))
+      .then(confirm => confirm || Promise.reject('User cancelled'))
       .then(() =>
         appsClient().uninstallApp(
           getAccount(),
@@ -259,6 +290,12 @@ export default {
   publish: {
     description: 'Publish this app',
     handler: () => {
+      const workspace = getWorkspace()
+      if (workspace === 'master') {
+        log.error(workspaceMasterMessage)
+        return Promise.resolve()
+      }
+
       log.debug('Starting to publish app')
       setSpinnerText('Publishing app...')
       startSpinner()
@@ -294,6 +331,12 @@ export default {
       description: 'Set a value',
       requiredArgs: ['app', 'field', 'value'],
       handler: async (app, field, value) => {
+        const workspace = getWorkspace()
+        if (workspace === 'master') {
+          log.error(workspaceMasterMessage)
+          return Promise.resolve()
+        }
+
         const patch = {}
         jp.value(patch, '$.' + field, value)
         const response = await appsClient().patchAppSettings(
@@ -306,6 +349,12 @@ export default {
       description: 'Unset a value',
       requiredArgs: ['app', 'field'],
       handler: async (app, field) => {
+        const workspace = getWorkspace()
+        if (workspace === 'master') {
+          log.error(workspaceMasterMessage)
+          return Promise.resolve()
+        }
+
         const patch = {}
         jp.value(patch, '$.' + field, null)
         const response = await appsClient().patchAppSettings(
